refactor: reorganize handler methods into separate files for better structure
